add changeRole handler call to changeRole modifier
